refactor: cleanup doc components
